chore: moved `delayAnimationFrames()` to separate file
